review software drivers

+++ software/iob_clint.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <stdint.h>
-
 #define MSIP_BASE 0
 #define MTIMECMP_BASE 16384
 #define MTIME_BASE 49144
@@ -12,8 +10,8 @@ void clint_init(int);
 
 void clint_set_timer(unsigned long long);
 void clint_set_timercmp(unsigned long long, int);
-void clint_set_msip(int, int);
+void clint_set_msip(unsigned long, int);
 
 unsigned long long clint_get_timer();
 unsigned long long clint_get_timercmp(int);
-unsigned int clint_get_msip(int);
+unsigned long clint_get_msip(int);
